% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_test_bot_config.R
\name{assert_data_cleaner_bot_config}
\alias{assert_data_cleaner_bot_config}
\title{Assert a \code{data_cleaner_bot_config}}
\usage{
assert_data_cleaner_bot_config(cfg)
}
\arguments{
\item{cfg}{an object to check}
}
\description{
Assert a \code{data_cleaner_bot_config}
}
\keyword{internal}
