% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_test_args.R
\name{assert_test_args}
\alias{assert_test_args}
\title{Assert test bash arguments}
\usage{
assert_test_args(args)
}
\arguments{
\item{args}{character vector to assert.}
}
\description{
Function to assert the arguments sent from bash to the production script.
}
\keyword{internal}
