% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{data_filter}
\alias{data_filter}
\title{Function for filtering out unwanted data}
\usage{
data_filter(dat, net_sales_over = NULL, tot_turnover_over = NULL,
  nbr_emp_over = NULL, growth_emp_over = NULL,
  growth_turnover_over = NULL, prof_marg_over = NULL,
  net_sales_under = NULL, tot_turnover_under = NULL,
  nbr_emp_under = NULL, growth_emp_under = NULL,
  growth_turnover_under = NULL, prof_marg_under = NULL, laen = NULL,
  city = NULL, sector = NULL)
}
\arguments{
\item{dat}{a dataset containg information about companies one wants to extract}

\item{laen}{parameter which one uses to filter the extisting data frame}

\item{city}{parameter which one uses to filter the extisting data frame}

\item{sector}{parameter which one uses to filter the extisting data frame}

\item{net_sales}{parameter which one uses to filter the extisting data frame}

\item{tot_turnover}{parameter which one uses to filter the extisting data frame}

\item{nbr_emp}{parameter which one uses to filter the extisting data frame}

\item{growth_emp}{parameter which one uses to filter the extisting data frame}

\item{growth_turnover}{parameter which one uses to filter the extisting data frame}

\item{prof_margh}{parameter which one uses to filter the extisting data frame}
}
\description{
Function for filtering out unwanted data
}
