% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_bot_config.R
\name{get_test_bot_config}
\alias{get_test_bot_config}
\title{Get data test config}
\usage{
get_test_bot_config(cfg_file)
}
\arguments{
\item{cfg_file}{config file path}
}
\value{
A \code{data_test_config} object.
}
\description{
Function to get and parse config
}
