% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{model_evaluation}
\alias{model_evaluation}
\title{Compute evaluation metrics for a given model}
\usage{
model_evaluation(test_model, test_data, cfg)
}
\arguments{
\item{test_model}{a \code{prospecting_leads} object for the validation set.}

\item{cfg}{a \code{prospecting_bot_config} object.}

\item{train_data}{a \code{train_data_set} object with the variable \code{validate_set} that indicates if a validate set or training set.}
}
\description{
Compute evaluation metrics for a given model
}
