% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_func}
\alias{predict_func}
\title{Predict best thingies by model and data}
\usage{
predict_func(test_model, data_set, cfg)
}
\arguments{
\item{test_model}{a \code{prospecting_model} object.}

\item{data_set}{a \code{company_data} object.}

\item{cfg}{a \code{prospecting_bot_config} object.}
}
\description{
A function to use predict on model used for testing
}
