% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_evaluation.R
\name{rank_evaluation}
\alias{rank_evaluation}
\title{Compute rank evaluation metrics}
\usage{
rank_evaluation(full_data, normalized = TRUE, fold = "validation_set",
  use_y = TRUE, FUN = mean, ...)
}
\arguments{
\item{full_data}{A \code{prospecting_leads} object with variables \code{company_id}, \code{score}, \code{true_value} and \code{validation_set}.}

\item{normalized}{Should the rank be normalized from 0 to 1?}

\item{fold}{For what fold should the rank be computed. \code{full_training_set} or \code{validation_set}.}

\item{use_y}{Use \code{y} variable to compute rank. If \code{FALSE}, use unique values in \code{sale_process_stage}.}

\item{FUN}{Function to compute the rank statistic for the 1:s.}

\item{...}{Further arguments to \code{FUN}.}
}
\description{
Compute rank evaluation metrics
}
\details{
This functions computes rank evaluations on the full \code{company_data}.
}
