% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Function to assist with reading data}
\usage{
read_data(cfg)
}
\arguments{
\item{cfg}{config containing file path to data}
}
\description{
Function to assist with reading data
}
