% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_test_args.R
\name{test_args}
\alias{test_args}
\title{Assert test bash arguments}
\usage{
test_args(args)
}
\arguments{
\item{args}{character vector to assert.}
}
\description{
If bash is used to call a function, assertion is of use since they will help one find
errors as early as possible in the process. This function is used to assert that the 
arguments from bash is correct.
}
