% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_test_bot_config.R
\name{test_bot_config}
\alias{test_bot_config}
\title{Create a \code{data_cleaner_bot_config} object}
\usage{
test_bot_config(cfg)
}
\arguments{
\item{cfg}{an \code{tbl_df} to convert to a \code{test_bot_config} object.}
}
\description{
Create a \code{data_cleaner_bot_config} object
}
