% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train model}
\usage{
train_model(data_set, cfg)
}
\arguments{
\item{data_set}{a training data set}

\item{cfg}{a \code{test_config} object.}
}
\value{
a \code{model} object
}
\description{
Train model
}
