% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{implemented_prospecting_models}
\alias{implemented_prospecting_models}
\alias{train_prospecting_model.log_reg}
\alias{train_prospecting_model.log_enet}
\alias{train_prospecting_model.svm_e1071}
\alias{train_prospecting_model.log_elastic_net}
\alias{train_prospecting_model.xgboost}
\title{Implemented models}
\usage{
implemented_prospecting_models()

train_prospecting_model.log_reg(data_set, cfg)

train_prospecting_model.log_enet(data_set, cfg, alpha, lambda = NULL)

train_prospecting_model.svm_e1071(data_set, cfg)

train_prospecting_model.log_elastic_net(data_set, cfg)

train_prospecting_model.xgboost(data_set, cfg)
}
\arguments{
\item{alpha}{Elastic net parameter.}

\item{lambda}{Penalty parameter for the elastic net.}
}
\description{
Implemented models
}
