% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Train test split}
\usage{
train_test_split(data, cfg)
}
\arguments{
\item{data}{data set to be used in train test split}

\item{cfg}{a test bot config}
}
\description{
Train test split
}
